% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{parse_os_help}
\alias{parse_os_help}
\title{Parse the arguments of a operating system call.}
\usage{
parse_os_help(os_response)
}
\arguments{
\item{os_response}{A character. The operating system response to a call.}
}
\value{
A character
}
\description{
Parse the arguments of a operating system call.
}
\examples{
\dontrun{
library(magrittr)
system2("gdalinfo", stdout = TRUE, stderr = TRUE) \%>\%
    parse_os_help()
}
}
\author{
Alber Sanchez, \email{alber.ipia@inpe.br}
}
